requirements.txt
streamlit
docxtpl
python-docx
mammoth